rag-pipeline/requirements.txt
litellm>=1.0.0
PyJWT>=2.6.0
redis>=4.5.0
requests>=2.28.0
psycopg2-binary>=2.9.5
python-multipart>=0.0.6
python-dotenv>=1.0.0
logging>=0.4.9 